requirements.txt
streamlit
streamlit-authenticator
streamlit-option-menu
PyYaml
matplotlib==3.5.2
pyparsing==2.4.7
htbuilder
sklearn
pandas
xgboost==1.6.1
seaborn



